feat: clippy

+++ src/drums/synth/kick.rs
@@ -6,11 +6,11 @@ pub fn process(params: &ElectricParams, pos: usize, vel: f32, sr: f32, phase: &m
     let t = pos as f32 / sr;
 
     // 振幅エンベロープ
-    let amp_env = SynthUtils::exp_env(t, params.decay.value() as f32);
+    let amp_env = SynthUtils::exp_env(t, params.decay.value());
 
     // ピッチエンベロープ: モダンなキックは最初の一瞬だけ超高域から落ちる
-    let f_base = params.freq.value() as f32;
-    let sweep_amount = params.sweep.value() as f32 * 8.0;
+    let f_base = params.freq.value();
+    let sweep_amount = params.sweep.value() * 8.0;
     let p_env = (-(t / 0.025)).exp(); // 非常に速いピッチ降下
     let freq = f_base * (1.0 + sweep_amount * p_env);
 
@@ -24,7 +24,7 @@ pub fn process(params: &ElectricParams, pos: usize, vel: f32, sr: f32, phase: &m
 
     // アタックノイズ
     let noise_env = (-(t / 0.004)).exp();
-    let noise = SynthUtils::stable_noise(pos) * noise_env * params.noise_level.value() as f32;
+    let noise = SynthUtils::stable_noise(pos) * noise_env * params.noise_level.value();
 
     (osc * amp_env + noise) * vel
 }
